# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 608a2fc1a3e049112324909d
  title: schoolmng
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/exam/findBy_student/{key}":
    get:
      x-skaffolder-id: 608a34cba3e0491123249156
      x-skaffolder-name: findBy_student
      x-skaffolder-id-resource: 608a33ada3e0491123249149
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_student
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_teacher/{key}":
    get:
      x-skaffolder-id: 608a350da3e0491123249164
      x-skaffolder-name: findBy_teacher
      x-skaffolder-id-resource: 608a33ada3e0491123249149
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_teacher
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
...
  "/student/{id}":
    delete:
      x-skaffolder-id: 608a34cba3e049112324915b
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 608a31daa3e0491123249134
      x-skaffolder-resource: student
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 608a34cba3e0491123249152
      x-skaffolder-name: get
      x-skaffolder-id-resource: 608a31daa3e0491123249134
      x-skaffolder-resource: student
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
    post:
      x-skaffolder-id: 608a34cba3e0491123249153
      x-skaffolder-name: update
      x-skaffolder-id-resource: 608a31daa3e0491123249134
      x-skaffolder-resource: student
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/student"
  "/student/findBy_courses/{key}":
    get:
      x-skaffolder-id: 608a3516a3e0491123249172
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 608a31daa3e0491123249134
      x-skaffolder-resource: student
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - student
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
...
  "/teacher/{id}":
    delete:
      x-skaffolder-id: 608a350ea3e0491123249167
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 608a3213a3e049112324913a
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 608a350da3e049112324915f
      x-skaffolder-name: get
      x-skaffolder-id-resource: 608a3213a3e049112324913a
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
    post:
      x-skaffolder-id: 608a350da3e0491123249162
      x-skaffolder-name: update
      x-skaffolder-id-resource: 608a3213a3e049112324913a
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/teacher"
  "/teacher/findBy_courses/{key}":
    get:
      x-skaffolder-id: 608a3515a3e0491123249170
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 608a3213a3e049112324913a
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - teacher
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teacher"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    exam:
      x-skaffolder-id: 608a33ada3e0491123249149
      x-skaffolder-id-db: 608a2fc1a3e049112324909e
      x-skaffolder-id-entity: 608a33ada3e0491123249146
      x-skaffolder-url: /exam
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        place:
          type: string
          x-skaffolder-id-attr: 608a33ada3e0491123249148
          x-skaffolder-type: String
          x-skaffolder-required: true
        score:
          type: number
          x-skaffolder-id-attr: 608a33ada3e0491123249147
          x-skaffolder-type: Number
          x-skaffolder-required: true
        validate:
          type: boolean
          x-skaffolder-id-attr: 608a3632a3e049112324917a
          x-skaffolder-type: Boolean
      x-skaffolder-relations:
        _student:
          x-skaffolder-id: 608a342ca3e049112324914b
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 608a33ada3e0491123249146
          x-skaffolder-ent2: 608a31daa3e0491123249130
        _teacher:
          x-skaffolder-id: 608a342ca3e049112324914d
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 608a33ada3e0491123249146
          x-skaffolder-ent2: 608a3213a3e0491123249137
      required:
        - _id
        - place
        - score
    student:
      x-skaffolder-id: 608a31daa3e0491123249134
      x-skaffolder-id-db: 608a2fc1a3e049112324909e
      x-skaffolder-id-entity: 608a31daa3e0491123249130
      x-skaffolder-url: /student
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        dob:
          type: integer
          x-skaffolder-id-attr: 608a31daa3e0491123249133
          x-skaffolder-type: Date
        name:
          type: string
          x-skaffolder-id-attr: 608a31daa3e0491123249131
          x-skaffolder-type: String
          x-skaffolder-required: true
        surname:
          type: string
          x-skaffolder-id-attr: 608a31daa3e0491123249132
          x-skaffolder-type: String
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 608a348da3e049112324914f
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 608a31daa3e0491123249130
          x-skaffolder-ent2: 608a3377a3e049112324913f
      required:
        - _id
        - name
    teacher:
      x-skaffolder-id: 608a3213a3e049112324913a
      x-skaffolder-id-db: 608a2fc1a3e049112324909e
      x-skaffolder-id-entity: 608a3213a3e0491123249137
      x-skaffolder-url: /teacher
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 608a3213a3e0491123249138
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 608a3213a3e0491123249139
          x-skaffolder-type: String
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 608a347da3e049112324914e
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 608a3213a3e0491123249137
          x-skaffolder-ent2: 608a3377a3e049112324913f
      required:
        - _id
...